chore: rename mapper
